+++ src/test/list.js
@@ -134,10 +134,12 @@ describe("Foxx service list", () => {
 
   describe("with a password file", () => {
     const user = "testuser";
+    const passwordFilePath = path.resolve(basePath, "passwordFile");
+    const passwd = fs.readFileSync(passwordFilePath, "utf-8");
     before(async () => {
       db.route("/_api/user").post({
         user,
-        passwd: "1234" // from fixtures/passwordFile
+        passwd
       });
       db.route(`/_api/user/${user}/database/_system`).put({ grant: "rw" });
     });
@@ -149,9 +151,8 @@ describe("Foxx service list", () => {
       }
     });
     it("should show information about the service", async () => {
-      const passwordFilePath = path.resolve(basePath, "passwordFile");
       const services = await foxx(
-        `list --username ${user} --passwordFile ${passwordFilePath}`,
+        `list --username ${user} --password-file ${passwordFilePath}`,
         true
       );
       expect(services).to.be.instanceOf(Array);