+++ src/test/scripts.js
@@ -84,10 +84,12 @@ describe("Foxx service scripts", () => {
 
   describe("with a password file", () => {
     const user = "testuser";
+    const passwordFilePath = path.resolve(basePath, "passwordFile");
+    const passwd = fs.readFileSync(passwordFilePath, "utf-8");
     before(async () => {
       db.route("/_api/user").post({
         user,
-        passwd: "1234" // from fixtures/passwordFile
+        passwd
       });
       db.route(`/_api/user/${user}/database/_system`).put({ grant: "rw" });
     });
@@ -99,9 +101,8 @@ describe("Foxx service scripts", () => {
       }
     });
     it("should all be listed", async () => {
-      const passwordFilePath = path.resolve(basePath, "passwordFile");
       const scripts = await foxx(
-        `scripts ${mount} --username ${user} --passwordFile ${passwordFilePath}`,
+        `scripts ${mount} --username ${user} --password-file ${passwordFilePath}`,
         true
       );
       expect(scripts).to.have.property("setup", "Setup");