+++ src/test/set-prod.js
@@ -102,10 +102,12 @@ describe("Foxx service production mode", () => {
 
   describe("with a password file", () => {
     const user = "testuser";
+    const passwordFilePath = path.resolve(basePath, "passwordFile");
+    const passwd = fs.readFileSync(passwordFilePath, "utf-8");
     before(async () => {
       db.route("/_api/user").post({
         user,
-        passwd: "1234" // from fixtures/passwordFile
+        passwd
       });
       db.route(`/_api/user/${user}/database/_system`).put({ grant: "rw" });
     });
@@ -117,11 +119,10 @@ describe("Foxx service production mode", () => {
       }
     });
     it("should be activated", async () => {
-      const passwordFilePath = path.resolve(basePath, "passwordFile");
       const infoBefore = await db.getService(mount);
       expect(infoBefore.development).to.equal(true);
       await foxx(
-        `set-prod --username ${user} --passwordFile ${passwordFilePath} ${mount}`
+        `set-prod --username ${user} --password-file ${passwordFilePath} ${mount}`
       );
       const infoAfter = await db.getService(mount);
       expect(infoAfter.development).to.equal(false);